RF: separated connections.py into news, sendusage, updates
