+++ psychopy/app/connections/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# Part of the PsychoPy library
+# Copyright (C) 2018 Jonathan Peirce
+# Distributed under the terms of the GNU General Public License (GPL).
+
+from .updates import *  # pylint: disable=W0401
+from .news import getNewsItems, NewsFrame  # pylint: disable=W0401
+from .sendusage import sendUsageStats
